+++ projects/example/src/app/app.component.ts
@@ -1,6 +1,6 @@
-import { Component } from '@angular/core';
-import { AppRatingComponent } from './rating.component';
-import { SmartTableConfig } from 'projects/smart-table/src/lib/smart-table/smart-table.types';
+import {Component} from '@angular/core';
+import {AppRatingComponent} from './rating.component';
+import {SmartTableConfig} from 'projects/smart-table/src/lib/smart-table/smart-table.types';
 
 @Component({
   selector: 'app-root',
@@ -15,14 +15,14 @@ export class AppComponent {
     component: AppRatingComponent
   }];
 
-  customConfiguration = {
+  customConfiguration: SmartTableConfig = {
     options: {
       persistTableConfig: true,
-      storageIdentifier: 'test-tc',
+      storageIdentifier: 'test-smart-table',
       translations: {
         moreFilters: 'Extra filters',
         export: 'Exporteer',
-        apply: 'Toepassen'
+        apply: 'Toepassen',
       }
     }
   };