+++ projects/smart-table/src/lib/indentifier.provider.ts
@@ -1,0 +1,3 @@
+import {InjectionToken} from '@angular/core';
+
+export const PROVIDE_ID = new InjectionToken('Provide-Identifier');