+++ projects/smart-table/src/lib/smart-table.module.ts
@@ -3,17 +3,20 @@ import {SmartTableComponent} from './smart-table/smart-table.component';
 import {CommonModule, DatePipe} from '@angular/common';
 import {HttpClientModule} from '@angular/common/http';
 import {ReactiveFormsModule} from '@angular/forms';
-import {TableModule} from '@acpaas-ui/ngx-components/table';
-import {ItemCounterModule, PaginationModule} from '@acpaas-ui/ngx-components/pagination';
-import {DatepickerModule, SearchFilterModule} from '@acpaas-ui/ngx-components/forms';
-import {FlyoutModule} from '@acpaas-ui/ngx-components/flyout';
-import {
-  DEFAULT_LOCALSTORAGE_CONFIG,
-  LOCALSTORAGE_CONFIG,
-  LocalstorageConfig,
-  LocalstorageModule
-} from '@acpaas-ui/ngx-components/localstorage';
+import {TableModule} from '@acpaas-ui/ngx-table';
+import {ItemCounterModule, PaginationModule} from '@acpaas-ui/ngx-pagination';
+import {DatepickerModule, SearchFilterModule} from '@acpaas-ui/ngx-forms';
+import {FlyoutModule} from '@acpaas-ui/ngx-flyout';
 import {components, services} from './index';
+import {LOCALSTORAGE_CONFIG, LocalstorageModule} from '@acpaas-ui/ngx-localstorage';
+import {IModuleConfig} from './smart-table/smart-table.types';
+import {PROVIDE_ID} from './indentifier.provider';
+import {TableFactory} from './services/table.factory';
+
+const defaultConfiguration: IModuleConfig = {
+  storageType: 'localStorage',
+  identifier: 'aui-smart-table'
+};
 
 @NgModule({
   declarations: [
@@ -29,11 +32,16 @@ import {components, services} from './index';
     HttpClientModule,
     FlyoutModule,
     SearchFilterModule,
-    LocalstorageModule
+    LocalstorageModule.forRoot(defaultConfiguration)
   ],
   providers: [
     DatePipe,
-    ...services
+    ...services,
+    TableFactory,
+    {
+      provide: PROVIDE_ID,
+      useValue: null
+    }
   ],
   exports: [
     SmartTableComponent,
@@ -41,16 +49,25 @@ import {components, services} from './index';
   ]
 })
 export class SmartTableModule {
-  static forRoot(
-    localstorageConfig: LocalstorageConfig = DEFAULT_LOCALSTORAGE_CONFIG
-  ): ModuleWithProviders {
+  static forRoot(localstorageConfig: IModuleConfig = defaultConfiguration): ModuleWithProviders {
     return {
       ngModule: SmartTableModule,
       providers: [
-        {provide: LOCALSTORAGE_CONFIG, useValue: localstorageConfig},
+        {
+          provide: PROVIDE_ID,
+          useValue: localstorageConfig.identifier
+        },
+        {
+          provide: LOCALSTORAGE_CONFIG,
+          useValue: {
+            storageType: localstorageConfig.storageType
+          }
+        },
         DatePipe,
-        ...services
+        ...services,
+        TableFactory
       ],
     };
   }
 }
+